Added doctest unit test framework
Also rewritten string tests in this library
