+++ SConstruct
@@ -539,6 +539,9 @@ if selected_platform in platform_list:
         env.Append(BUILDERS = { 'GLES2_GLSL' : env.Builder(action=run_in_subprocess(gles_builders.build_gles2_headers), suffix='glsl.gen.h', src_suffix='.glsl')})
         env.Append(BUILDERS = { 'RD_GLSL' : env.Builder(action=run_in_subprocess(gles_builders.build_rd_headers), suffix='glsl.gen.h', src_suffix='.glsl')})
 
+    # enable test framework globally and inform it of configuration method
+    env.Append(CPPDEFINES=['DOCTEST_CONFIG_IMPLEMENT'])
+
     scons_cache_path = os.environ.get("SCONS_CACHE")
     if scons_cache_path != None:
         CacheDir(scons_cache_path)
