+++ main/SCsub
@@ -8,6 +8,7 @@ import main_builders
 env.main_sources = []
 
 env.add_source_files(env.main_sources, "*.cpp")
+env.add_source_files(env.main_sources, "tests/*.cpp")
 
 # Order matters here. Higher index controller database files write on top of lower index database files.
 controller_databases = ["#main/gamecontrollerdb_204.txt", "#main/gamecontrollerdb_205.txt", "#main/gamecontrollerdb.txt", "#main/godotcontrollerdb.txt"]
@@ -27,9 +28,6 @@ env.CommandNoCache("#main/splash_editor.gen.h", "#main/splash_editor.png", run_i
 
 env.Depends("#main/app_icon.gen.h", "#main/app_icon.png")
 env.CommandNoCache("#main/app_icon.gen.h", "#main/app_icon.png", run_in_subprocess(main_builders.make_app_icon))
-
-if env["tools"]:
-    SConscript('tests/SCsub')
 
 lib = env.add_library("main", env.main_sources)
 env.Prepend(LIBS=[lib])
