+++ main/tests/test_main.cpp
@@ -33,6 +33,8 @@
 
 #ifdef DEBUG_ENABLED
 
+#include <thirdparty/doctest/doctest.h>
+
 #include "test_astar.h"
 #include "test_gdscript.h"
 #include "test_gui.h"
@@ -48,7 +50,7 @@
 const char **tests_get_names() {
 
 	static const char *test_names[] = {
-		"string",
+		"doctest",
 		"math",
 		"physics",
 		"physics_2d",
@@ -69,10 +71,15 @@ const char **tests_get_names() {
 }
 
 MainLoop *test_main(String p_test, const List<String> &p_args) {
+	doctest::Context test_context;
 
-	if (p_test == "string") {
-
-		return TestString::test();
+	if (p_test == "doctest") {
+		// tests should be ordered by name for humans
+		test_context.setOption("order-by", "name");
+		test_context.setOption("abort-after", 5);
+		test_context.setOption("no-breaks", true);
+		test_context.run(); // run tests
+		return NULL;
 	}
 
 	if (p_test == "math") {
